add default generic service

+++ spring-boot-eval-service/spring-boot-eval-service-interface/src/main/java/com/smec/eis/example/springbooteval/service/QueryDslSupport.java
@@ -9,6 +9,6 @@ public interface QueryDslSupport<T> {
 
     List<T> query(EntityPath<T> from, Predicate predicate);
 
-    List<T> query(EntityPath<T> from, Predicate predicate, long offset, long limit);
+    List<T> query(EntityPath<T> from, Predicate predicate, int offset, int limit);
 
 }
